Update Win32App_PSADT_PreLaunch.ps1
Defining the below in Get-CIMInstance was causing failures of the command on new autopilot deployed devices, it would generate an error.  There is not really a need to define it as it is run on the local computer, so I removed it, and that cleared the errors and it works the same as before.  Something to do the instance possibly not having a PSComputerName defined, that was only difference I saw between a computer that worked and one that didn't.

-ComputerName $env:computername